requirements.txt
# Core Dependencies
langgraph==0.2.16
langchain==0.2.16
langchain-community==0.2.16
langchain-google-genai==1.0.10

# Data Validation & Processing
pydantic==2.8.2
pydantic-settings==2.4.0

# Configuration & Environment
python-dotenv==1.0.1
pyyaml==6.0.1

# HTTP & API Clients
requests==2.32.3
httpx==0.27.0

# Google Services
google-generativeai==0.7.2
gspread==6.1.2
google-auth==2.34.0
google-auth-oauthlib==1.2.1

# Email Services
sendgrid==6.11.0

# Database (Optional)
sqlalchemy==2.0.31

# Logging & Monitoring
colorlog==6.8.2

# Utilities
python-dateutil==2.9.0
pytz==2024.1

# Testing
pytest==8.3.2
pytest-asyncio==0.23.8